+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.2] - 2026-02-07
+
+### Added
+- **Infinite Scroll**: Added infinite scroll to the Watch List page to improve performance when loading a large number of channels.
+- **Default Filter**: Added option to set default filter for the Watch List page.   
+
+### Fixed
+- **Tags**: Fixed issue with tags capitalization in the Watch List page.
+
+### Docs
+- **Organization**: organized roadmap docs by release tag folders.
+
+
 ## [0.5.1] - 2026-02-06
 
 ### Docs